+++ App.tsx
@@ -2,9 +2,12 @@ import * as React from 'react'
 import { Provider } from 'react-redux'
 import { store } from './src/store/store'
 import { AppNavigation } from './App.navigation'
+import { NavigationContainer } from '@react-navigation/native'
 
 export const App = (): JSX.Element => (
   <Provider store={store}>
-    <AppNavigation />
+    <NavigationContainer>
+      <AppNavigation />
+    </NavigationContainer>
   </Provider>
 )